replaced the initial code with a simpler one. Expect future problem?
I don't understand the previous implementation of the waveform creation in case of step exec_rule. There was a mix in previous and current points, which resulted in additional (and fake) points in TOPHYB and FTOPHYB (and wrongly located...), and so a bad detection. Hope the present modification will not create edge effects...